+++ core/modules/moddiscountrules.class.php
@@ -73,7 +73,7 @@ class moddiscountrules extends DolibarrModules
 
 		// Possible values for version are: 'development', 'experimental', 'dolibarr', 'dolibarr_deprecated' or a version string like 'x.y.z'
 
-		$this->version = '2.14.3';
+		$this->version = '2.15.0';
 
 		// Key used in llx_const table to save module status enabled/disabled (where discountrules is value of property name of module in uppercase)
 		$this->const_name = 'MAIN_MODULE_'.strtoupper($this->name);
@@ -126,7 +126,7 @@ class moddiscountrules extends DolibarrModules
 		$this->conflictwith = array();	// List of module class names as string this module is in conflict with
 		$this->phpmin = array(5,6);					// Minimum version of PHP required by module
 		$this->need_dolibarr_version = array(10,0);	// Minimum version of Dolibarr required by module
-		$this->langfiles = array("discountrules@discountrules");
+		$this->langfiles = array("discountrules@discountrules","importdiscountrules@discountrules");
 		$this->warnings_activation = array();                     // Warning to show when we activate module. array('always'='text') or array('FR'='textfr','ES'='textes'...)
 		$this->warnings_activation_ext = array();                 // Warning to show when we activate an external module. array('always'='text') or array('FR'='textfr','ES'='textes'...)
 
@@ -242,6 +242,8 @@ class moddiscountrules extends DolibarrModules
 		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->discountrules->level1->level2)
 
 
+
+
 		// Main menu entries
 		$this->menu = array();			// List of menus to add
 		$r=0;
@@ -332,6 +334,26 @@ class moddiscountrules extends DolibarrModules
             'user'=>0
 		);
 
+		//Menu  import discount Rules
+		$r++;
+		$this->menu[$r] = array(
+			'fk_menu'=>'fk_mainmenu=tools,fk_leftmenu=import',		    // Use 'fk_mainmenu=xxx' or 'fk_mainmenu=xxx,fk_leftmenu=yyy' where xxx is mainmenucode and yyy is a leftmenucode
+			'type'=>'left',
+			'titre'=>'idrImportDiscountRules',
+			'prefix' => img_picto('', $this->picto, 'class="paddingright pictofixedwidth valignmiddle"'),
+			'mainmenu'=>'importdiscountrules',
+			'leftmenu'=>'importdiscountrules_left',
+
+			'url'=>'/discountrules/discount_rules_import.php?datatoimport=importdiscountrules&mainmenu=tools',
+			'langs'=>'importdiscountrules@discountrules', // Lang file to use (without .lang) by module. File must be in langs/code_CODE/ directory.
+			'position'=>1000 + $r,
+			'enabled'=>'$conf->discountrules->enabled', // Define condition to show or hide menu entry. Use '$conf->importdiscountrules->enabled' if entry must be visible if module is enabled.
+			'perms'=>'$user->rights->discountrules->create',			                // Use 'perms'=>'$user->rights->cliaufildesmatieres->level1->level2' if you want your menu with a permission rules
+			'target'=>'',
+			'user'=>0, // 0=Menu for internal users, 1=external users, 2=both
+		);
+
+
         $r++;
 
 
@@ -354,6 +376,68 @@ class moddiscountrules extends DolibarrModules
 		// $this->export_sql_order[$r] .=' ORDER BY t.ref';
 		// $r++;
 		END MODULEBUILDER EXPORT MYOBJECT */
+
+
+
+		// Imports profiles provided by this module
+		$r = 1;
+		/* BEGIN MODULEBUILDER IMPORT MYOBJECT */
+
+		$this->import_code[$r] = $this->rights_class.'_'.$r;
+		$this->import_label[$r] = "discountrules"; // Translation key
+		$this->import_icon[$r] = $this->picto;
+		// for example csv file
+		$this->import_fields_array[$r] = array(
+			"label" 					=> "label",
+			"fk_project" 				=> "refProject",
+			"fk_product" 				=> "refProduct",
+			"fk_company" 				=> "refCompany",
+			"fk_country" 				=> "refCountry",
+			"priority_rank" 			=>"priorityRank",
+			"fk_c_typent" 				=> "cTypeEnt",
+
+
+
+			"all_category_product" 		=>"allCategoryProduct",
+			"all_category_company" 		=>"allCategoryCompany",
+
+			"reduction" 				=> "reduction",
+			"from_quantity" 			=> "fromQuantity",
+			"product_price" 			=> "productPrice",
+			"product_reduction_amount" => "productReductionAmount",
+			"date_from" 				=>"dateFrom",
+			"date_to" 					=>"dateTo",
+			"activation" 				=>"activation",
+
+
+		);
+
+		//@todo exemple à remplir
+		$this->import_examplevalues_array[$r] = array(
+			"label" 					=> "ligne Exemple",
+
+			"fk_project" 				=> "PJ2201-0001",
+			"fk_product" 				=> "PRODUIT_IMPORT_01",
+			"fk_company" 				=> "KEVIN",
+			"fk_country" 				=> "code pays. ex :  US",
+			"priority_rank" 			=>"vide ou 0  si pas de priorité sinon numérique entre 1 et 5",
+			"fk_c_typent" 				=> "cTypeEnt",
+
+
+			"all_category_product" 		=>"vide pour toutes les catégories sinon liste des ref séparées par des virgules. ex : TCP01,TCP02",
+			"all_category_company" 		=>"vide pour toutes les catégories sinon liste des ref séparées par des virgules. ex : TCP01,TCP02",
+
+			"from_quantity" 			=> "numérique",
+			"product_price" 			=> "numérique",
+			"product_reduction_amount" => "5",
+			"reduction" 				=> "10",
+			"date_from" 				=>"date au format jj/mm/yyyy",
+			"date_to" 					=>"date au format jj/mm/yyyy",
+			"activation" 				=>"vide/0 pour désactiver 1 pour activer",
+
+			);
+		/* END MODULEBUILDER IMPORT MYOBJECT */
+
 
 	}