+++ ChangeLog.md
@@ -6,6 +6,7 @@
 
 ## 2.14
 
+- FIX : Popin compare country field in description if activated *05/01/2022* - 2.14.1
 - NEW : Last price search in document configuration *12/12/2021* - 2.14.0  
   The search in document feature finds the last price applied to a customer  
   A conf enables the module to search the best price instead of the last price