+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## Not Released
 
-
+- NEW : Import des règles de prix  *12/01/2022* - 2.15.0
 
 ## 2.14